⚙️ refactor(prisma): remove PrismaActionArgsWeaver
